tests/wpt/web-platform-tests/IndexedDB/idbfactory_open4.htm

<!DOCTYPE html>
<title>IDBFactory.open() - new database has default version</title>
<link rel="author" title="Microsoft" href="http://www.microsoft.com">
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=support.js></script>

<script>
    var open_rq = createdb(async_test(), document.location + '-database_name');

    open_rq.onupgradeneeded = function(e) {
        assert_equals(e.target.result.version, 1, "db.version");
    };
    open_rq.onsuccess = function(e) {
        assert_equals(e.target.result.version, 1, "db.version");
        this.done();
    };
</script>

<div id=log></div>
